+++ src/main/java/windeath44/server/memorial/domain/presentation/dto/global/ResponseDto.java
@@ -1,7 +1,7 @@
 package windeath44.server.memorial.domain.presentation.dto.global;
 
 public record ResponseDto(
-        Long status,
+        Integer status,
         String message,
         Object data
 ) {
